move request purpose validation to a seperate file
