+++ app/templates/request_purpose.hbs
@@ -23,7 +23,7 @@
             </section>
             <div class="row inner-section">
               <section class="main-section request-purpose medium-10 medium-offset-1 large-10 large-offset-1 small-10 small-offset-1 columns">
-              {{#validate-appointment}}
+              {{#validate-request-purpose}}
                 {{#validatable-form onSubmit=(action 'createOrderWithRequestPurpose') as |form|}}
                   <div class="small-12 columns">
                     {{t 'order.request_purpose.purpose'}}
@@ -123,7 +123,7 @@
                     {{/if}}
                   </div>
                 {{/validatable-form}}
-              {{/validate-appointment}}
+              {{/validate-request-purpose}}
             </section>
             </div>
         </div>
